Bugfix/v0.1.6

+++ products/ASC.Files/Client/src/components/pages/Home/Section/Body/FilesRow/SimpleFilesRow.js
@@ -14,6 +14,7 @@ import { FileAction, AppServerConfig } from "@appserver/common/constants";
 import copy from "copy-to-clipboard";
 import config from "../../../../../../../package.json";
 import { combineUrl } from "@appserver/common/utils";
+import { createSelectable } from "react-selectable-fast";
 
 const StyledSimpleFilesRow = styled(Row)`
   margin-top: -2px;
@@ -62,7 +63,7 @@ const EncryptedFileIcon = styled.div`
 
 const svgLoader = () => <div style={{ width: "24px" }}></div>;
 
-const SimpleFilesRow = (props) => {
+const SimpleFilesRow = createSelectable((props) => {
   const {
     t,
     item,
@@ -110,6 +111,8 @@ const SimpleFilesRow = (props) => {
     setTooltipPosition,
     setDownloadDialogVisible,
     downloadAction,
+    confirmDelete,
+    setDeleteDialogVisible,
   } = props;
 
   const {
@@ -269,17 +272,21 @@ const SimpleFilesRow = (props) => {
       return;
     }
 
-    const translations = {
-      deleteOperation: t("DeleteOperation"),
-    };
+    if (confirmDelete) {
+      setDeleteDialogVisible(true);
+    } else {
+      const translations = {
+        deleteOperation: t("DeleteOperation"),
+      };
 
-    fileExst || contentLength
-      ? deleteFileAction(id, folderId, translations)
-          .then(() => toastr.success(t("FileRemoved")))
-          .catch((err) => toastr.error(err))
-      : deleteFolderAction(id, parentId, translations)
-          .then(() => toastr.success(t("FolderRemoved")))
-          .catch((err) => toastr.error(err));
+      fileExst || contentLength
+        ? deleteFileAction(id, folderId, translations)
+            .then(() => toastr.success(t("FileRemoved")))
+            .catch((err) => toastr.error(err))
+        : deleteFolderAction(id, parentId, translations)
+            .then(() => toastr.success(t("FolderRemoved")))
+            .catch((err) => toastr.error(err));
+    }
   };
 
   const rowContextClick = () => {
@@ -288,7 +295,6 @@ const SimpleFilesRow = (props) => {
 
   const getFilesContextOptions = useCallback(() => {
     const isSharable = access !== 1 && access !== 0;
-
     return contextOptions.map((option) => {
       switch (option) {
         case "open":
@@ -417,7 +423,7 @@ const SimpleFilesRow = (props) => {
             onClick: onClickDownloadAs,
             disabled: false,
           };
-        case "move":
+        case "move-to":
           return {
             key: option,
             label: t("MoveTo"),
@@ -433,7 +439,7 @@ const SimpleFilesRow = (props) => {
             onClick: onMoveAction,
             disabled: false,
           };
-        case "copy":
+        case "copy-to":
           return {
             key: option,
             label: t("Copy"),
@@ -441,7 +447,7 @@ const SimpleFilesRow = (props) => {
             onClick: onCopyAction,
             disabled: false,
           };
-        case "duplicate":
+        case "copy":
           return {
             key: option,
             label: t("Duplicate"),
@@ -547,7 +553,8 @@ const SimpleFilesRow = (props) => {
   const element = getItemIcon(isEdit || id <= 0);
   const displayShareButton = isMobile ? "26px" : !canShare ? "38px" : "96px";
   let className = isDragging ? " droppable" : "";
-  if (draggable) className += " draggable";
+  if (draggable) className += " draggable not-selectable";
+
   let value = fileExst || contentLength ? `file_${id}` : `folder_${id}`;
   value += draggable ? "_draggable" : "";
 
@@ -557,32 +564,35 @@ const SimpleFilesRow = (props) => {
       : getSharedButton(shared);
 
   return (
-    <DragAndDrop
-      value={value}
-      className={className}
-      onDrop={onDrop}
-      onMouseDown={onMouseDown}
-      dragging={dragging && isDragging}
-      {...contextOptionsProps}
-    >
-      <StyledSimpleFilesRow
-        key={id}
-        data={item}
-        element={element}
-        sectionWidth={sectionWidth}
-        contentElement={sharedButton}
-        onSelect={onContentRowSelect}
-        rowContextClick={rowContextClick}
-        isPrivacy={isPrivacy}
-        {...checkedProps}
+    <div ref={props.selectableRef}>
+      <DragAndDrop
+        data-title={item.title}
+        value={value}
+        className={className}
+        onDrop={onDrop}
+        onMouseDown={onMouseDown}
+        dragging={dragging && isDragging}
         {...contextOptionsProps}
-        contextButtonSpacerWidth={displayShareButton}
       >
-        <FilesRowContent item={item} sectionWidth={sectionWidth} />
-      </StyledSimpleFilesRow>
-    </DragAndDrop>
+        <StyledSimpleFilesRow
+          key={id}
+          data={item}
+          element={element}
+          sectionWidth={sectionWidth}
+          contentElement={sharedButton}
+          onSelect={onContentRowSelect}
+          rowContextClick={rowContextClick}
+          isPrivacy={isPrivacy}
+          {...checkedProps}
+          {...contextOptionsProps}
+          contextButtonSpacerWidth={displayShareButton}
+        >
+          <FilesRowContent item={item} sectionWidth={sectionWidth} />
+        </StyledSimpleFilesRow>
+      </DragAndDrop>
+    </div>
   );
-};
+});
 
 export default inject(
   (
@@ -596,6 +606,7 @@ export default inject(
       filesActionsStore,
       mediaViewerDataStore,
       uploadDataStore,
+      settingsStore,
     },
     { item }
   ) => {
@@ -611,6 +622,7 @@ export default inject(
       setMoveToPanelVisible,
       setCopyPanelVisible,
       setDownloadDialogVisible,
+      setDeleteDialogVisible,
     } = dialogsStore;
 
     const {
@@ -622,6 +634,7 @@ export default inject(
       setDragging,
       setStartDrag,
       setTooltipPosition,
+      isFileSelected,
     } = filesStore;
 
     const { isRootFolder, id: selectedFolderId } = selectedFolderStore;
@@ -666,7 +679,7 @@ export default inject(
       isRecycleBin: isRecycleBinFolder,
       isRootFolder,
       canShare,
-      checked: selection.some((el) => el.id === item.id),
+      checked: isFileSelected(item.id, item.parentId),
       isFolder,
       draggable,
       isItemsSelected: !!selection.length,
@@ -701,6 +714,8 @@ export default inject(
       onSelectItem,
       setTooltipPosition,
       downloadAction,
+      confirmDelete: settingsStore.confirmDelete,
+      setDeleteDialogVisible,
     };
   }
 )(withTranslation("Home")(observer(withRouter(SimpleFilesRow))));
